backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
aiofiles==24.1.0
psutil==6.0.0
watchdog==4.0.1
pypdf>=4.0.0,<6.0.0
python-docx==1.1.2
python-multipart==0.0.12
